refactor(otp): organise logic in routes & services layer
The code does look cleaner now
